services/video/workers/svdBridge/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
diffusers==0.24.0
transformers==4.35.0
accelerate==0.24.1
opencv-python==4.8.1.78
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
requests==2.31.0
